add formatting and loader
